Fixed warnings
There were several warnings that compiling would give.  Those have been fixed.

+++ src/EasyStarterKitIRremoteInt.h
@@ -192,7 +192,8 @@ typedef struct {
   uint8_t blinkflag;         // TRUE to enable blinking of pin 13 on IR processing
   unsigned int timer;     // state timer, counts 50uS ticks.
   unsigned int rawbuf[RAWBUF]; // raw data
-  uint8_t rawlen;         // counter of entries in rawbuf
+  //uint8_t rawlen;         // counter of entries in rawbuf
+  int rawlen;         // counter of entries in rawbuf 		//GAFFORD 3-23-2023
 } 
 irparams_t;
